+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to
 
 ## [Unreleased]
 
+- 401 Unauthorized for authentication errors
+
 ## [0.48.0] - 2026-05-28
 
 ### Added